requirements.txt
fastapi
uvicorn
pymongo[srv]
passlib[bcrypt]
python-multipart
python-dotenv
dnspython
razorpay
pydantic[email]
email-validator